Bump to version 1.4b - Several improvements
Added Unsharp Mask, Normalize, Auto-Level, Trim, and Rotate switches. Added support for HDRi (64bit Color) files, including the infrared channel. Not NegFix can automatically make a mask out of infrared data and save it as alpha channel inside the tiff file. This functionality has not been added to the black and white branches and the branches requiring older versions of ImageMagick. This functionality has a bug that prevents it to work nicely with the newly added Trim and USM.